Fixed all the error issues.
