--- 09. cli-and-http-request/clihelper/Structs/WriteUsingLogger.go
@@ -1,20 +1,0 @@
-package clihelper
-
-type WriteUsingLogger struct {
-	WriteUsingTheLogger(cliBindingProperties CliBindingProperties,
-		line string)
-}
-
-import (
-. "github.com/nafrose/exploring/clirunner/clihelper/Interfaces"
-. "github.com/nafrose/exploring/clirunner/clihelper/Structs"
-)
-
-func WritUsingLoggers(
-	writers []Writer,
-	wc WriterConfiguration,
-	line string) {
-	for _, writer := range writers {
-		writer.Writer(wc, line)
-	}
-}
