+++ 09. cli-and-http-request/clihelper/WriteUsingOutputLoggers.go
@@ -2,7 +2,9 @@ package clihelper
 
 import . "github.com/nafrose/exploring/clirunner/clihelper/Structs"
 
-func WriteUsingOutputLoggers(cliBindingProperties CliBindingProperties, line string) {
+func WriteUsingOutputLoggers(
+	cliBindingProperties CliBindingProperties,
+	line string) {
 	writers := cliBindingProperties.WritersCollection.OutputLoggers
 	wc := cliBindingProperties.WritersCollection.WriterConfiguration
 	WritUsingLoggers(writers, wc, line)
